requirements.txt
fastapi
uvicorn
python-multipart
psycopg2
python-dotenv